requirements.txt
streamlit==1.13.0
yfinance==0.2.4
plotly-express==0.4.1